requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
gspread==6.1.4
google-auth==2.35.0
openpyxl==3.1.5
